# feat: Improve encoding/decoding performance for ASCII strings