Since in practice most Package URLs don't have any percent encoded characters, this PR optimizes the `percentEncode/percentDecode` methods by:

- Always testing if encoding/decoding is needed. If it is not needed, the methods return the argument unmodified.
- Improving the performance of `isUnreserved`, since profiling shows that this is the most important method for the performance of `percentEncode`.