@ppkarwasz 2025-03-23
## Benchmark before

| Benchmark                          | (nonAsciiProb) | Mode | Cnt | Score    | Error     | Units |
|------------------------------------|----------------|------|-----|----------|-----------|-------|
| StringUtilBenchmark.baseline       | 0              | avgt | 5   | 35.683   | ± 0.134   | us/op |
| StringUtilBenchmark.baseline       | 0.1            | avgt | 5   | 658.929  | ± 3.496   | us/op |
| StringUtilBenchmark.baseline       | 0.5            | avgt | 5   | 1965.363 | ± 47.918  | us/op |
| StringUtilBenchmark.percentDecode  | 0              | avgt | 5   | 142.311  | ± 37.302  | us/op |
| StringUtilBenchmark.percentDecode  | 0.1            | avgt | 5   | 930.824  | ± 23.453  | us/op |
| StringUtilBenchmark.percentDecode  | 0.5            | avgt | 5   | 2614.854 | ± 147.601 | us/op |
| StringUtilBenchmark.percentEncode  | 0              | avgt | 5   | 870.956  | ± 29.092  | us/op |
| StringUtilBenchmark.percentEncode  | 0.1            | avgt | 5   | 1844.124 | ± 62.144  | us/op |
| StringUtilBenchmark.percentEncode  | 0.5            | avgt | 5   | 3684.842 | ± 61.788  | us/op |
| StringUtilBenchmark.toLowerCase    | 0              | avgt | 5   | 103.010  | ± 3.429   | us/op |
| StringUtilBenchmark.toLowerCase    | 0.1            | avgt | 5   | 119.274  | ± 1.037   | us/op |
| StringUtilBenchmark.toLowerCase    | 0.5            | avgt | 5   | 111.013  | ± 3.908   | us/op |
| StringUtilBenchmark.toLowerCaseJre | 0              | avgt | 5   | 729.618  | ± 9.626   | us/op |
| StringUtilBenchmark.toLowerCaseJre | 0.1            | avgt | 5   | 981.569  | ± 44.689  | us/op |
| StringUtilBenchmark.toLowerCaseJre | 0.5            | avgt | 5   | 1472.047 | ± 32.888  | us/op |

## Benchmark after

| Benchmark                          | (nonAsciiProb) | Mode | Cnt | Score    | Error     | Units |
|------------------------------------|----------------|------|-----|----------|-----------|-------|
| StringUtilBenchmark.baseline       | 0              | avgt | 5   | 35.997   | ± 0.989   | us/op |
| StringUtilBenchmark.baseline       | 0.1            | avgt | 5   | 649.834  | ± 13.729  | us/op |
| StringUtilBenchmark.baseline       | 0.5            | avgt | 5   | 2112.914 | ± 37.607  | us/op |
| StringUtilBenchmark.percentDecode  | 0              | avgt | 5   | 4.110    | ± 0.091   | us/op |
| StringUtilBenchmark.percentDecode  | 0.1            | avgt | 5   | 875.127  | ± 30.614  | us/op |
| StringUtilBenchmark.percentDecode  | 0.5            | avgt | 5   | 2255.278 | ± 32.528  | us/op |
| StringUtilBenchmark.percentEncode  | 0              | avgt | 5   | 80.203   | ± 1.481   | us/op |
| StringUtilBenchmark.percentEncode  | 0.1            | avgt | 5   | 939.128  | ± 45.859  | us/op |
| StringUtilBenchmark.percentEncode  | 0.5            | avgt | 5   | 3351.013 | ± 109.419 | us/op |
| StringUtilBenchmark.toLowerCase    | 0              | avgt | 5   | 104.420  | ± 3.729   | us/op |
| StringUtilBenchmark.toLowerCase    | 0.1            | avgt | 5   | 119.050  | ± 1.828   | us/op |
| StringUtilBenchmark.toLowerCase    | 0.5            | avgt | 5   | 110.732  | ± 0.130   | us/op |
| StringUtilBenchmark.toLowerCaseJre | 0              | avgt | 5   | 717.085  | ± 45.253  | us/op |
| StringUtilBenchmark.toLowerCaseJre | 0.1            | avgt | 5   | 963.733  | ± 9.452   | us/op |
| StringUtilBenchmark.toLowerCaseJre | 0.5            | avgt | 5   | 1436.714 | ± 42.871  | us/op |